requirements.txt
django==5.2.10
pyclamd==0.4.0
PyPDF2==3.0.1
pillow==12.1.0
python-magic-bin==0.4.14; platform_system == "Windows"
python-magic==0.4.27; platform_system != "Windows"
